Allow specifying custom Z_PREFIX_ instead of "z_"
Instead of allowing building 2 versions of the library, one without any
prefix and another one with "z_" prefix, make it possible to build as
many of variants as needed to allow creating really unique zlib builds
not clashing with anything else.

Do this by using Z_PREFIX_ macro as the prefix if it is defined. The
intended usage is to pass -DZ_PREIFX_=mylib_embedding_zlib_ when
building zlib embedded into some other library, so that linking with
this library later doesn't conflict with using any other builds of zlib
(whether standalone or embedded by a third library).